+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Change Log
 
-All notable changes to the IntelliTest extension will be documented in this file.
+All notable changes to the Debuggo extension will be documented in this file.
+
+## [0.0.2] - 2026-05-09
+
+### Changed
+
+- Extension renamed from IntelliTest to **Debuggo** (marketplace id: `publisher.debuggo`). Settings key is now `debuggo.backendUrl`; sidebar view id is `debuggoView`.
 
 ## [0.0.1] - 2026-05-09